change route name and fix routes
