+++ src/pages/node/[title]/[id].tsx
@@ -79,7 +79,7 @@ const NodePage: NextPage<Props> = ({ node, keywords, createdStr, updatedStr }) =
     const userAuthObj = auth?.currentUser;
     if (userAuthObj !== null && node?.title && node?.id) {
       router.push({
-        pathname: ROUTES.dashboard,
+        pathname: ROUTES.notebook,
         query: { nodeId: node?.id },
       });
       return;
